+++ Sources/Sentry/Public/SentryOptions.h
@@ -93,7 +93,7 @@ NS_SWIFT_NAME(Options)
 /**
  * Array of integrations to install.
  */
-@property (nonatomic, copy) NSArray<NSString *> *_Nullable integrations;
+@property (nullable, nonatomic, copy) NSArray<NSString *> *integrations;
 
 /**
  * Array of default integrations. Will be used if integrations are nil
@@ -147,6 +147,20 @@ NS_SWIFT_NAME(Options)
  * address.
  */
 @property (nonatomic, assign) BOOL sendDefaultPii;
+
+/**
+ * Indicates the percentage of the tracing data that is collected.
+ * Setting this to 0 or NIL discards all trace data, 1.0 collects all trace data,
+ * 0.01 collects 1% of all trace data.
+ */
+@property (nullable, nonatomic, strong) NSNumber *tracesSampleRate;
+
+/**
+ * A callback to a user defined traces sampler function.
+ * Returning 0 or NIL discards all trace data, 1.0 collects all trace data,
+ * 0.01 collects 1% of all trace data.
+ */
+@property (nullable, nonatomic) SentryTracesSamplerCallback tracesSampler;
 
 @end